IOS-4784 Refactor QR code scanner into reusable view modifier
